People need to know, this can prevent a nightmare
This what can happen if a person didn't notice the first 'zshrc':

if a person has installed dirjump for bash and zsh,
when uninstalling he run this command to unsource dirjump from bash:
    grep -Fxv "source ~/.config/dirjump/dirjump" ~/.bashrc > temp; mv temp ~/.bashrc

when he want to unsource dirjump from zsh, instead of copying this command:
    grep -Fxv "source ~/.config/dirjump/dirjump" ~/.zshrc > temp; mv temp ~/.zshrc

he might think that it is faster to just press the up arrow(get the last command) and change the last bashrc to zshrc.
which will result the following:
    grep -Fxv "source ~/.config/dirjump/dirjump" ~/.zshrc > temp; mv temp ~/.bashrc

That command is so evil. Don't ever you try it unless you have git initialized in your home directory.